[Feat] 휴식시간 담당 구조체 RestManager 정의
